EDO-256: move teams-edit to shared directory
