+++ src/main/webapp/app/teams/teams.module.ts
@@ -16,7 +16,7 @@ import { SkillDetailsInfoComponent } from 'app/shared/skill-details/skill-detail
 import { SkillDetailsCommentsComponent } from 'app/shared/skill-details/skill-details-comments.component';
 import { SkillDetailsRatingComponent } from 'app/teams/skill-details/skill-details-rating/skill-details-rating.component';
 import { AllCommentsResolve, AllSkillsResolve, DojoModelResolve, SkillResolve } from 'app/shared/common.resolver';
-import { TeamsEditComponent } from 'app/teams/teams-edit.component';
+import { TeamsEditComponent } from 'app/shared/teams/teams-edit.component';
 
 @NgModule({
     imports: [TeamdojoSharedModule, RouterModule.forChild(TEAMS_ROUTES), NgbModule],
